fix(incr9): move ErrorJSONResponse and JSONResponse to handler_api.go
