+++ internal/http/handler/handler_api.go
@@ -50,3 +50,27 @@ func (sh Handler) PostAPI(w http.ResponseWriter, r *http.Request) {
 		Result: shortURL,
 	})
 }
+
+func ErrorJSONResponse(w http.ResponseWriter, code int, error string, message string) {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(code)
+	json.NewEncoder(w).Encode(ErrorResponse{Code: code, Error: error, Detail: message})
+}
+
+func JSONResponse(w http.ResponseWriter, statusCode int, payload interface{}) {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(statusCode)
+
+	if payload == nil {
+		return
+	}
+
+	data, err := json.Marshal(payload)
+
+	if err != nil {
+		ErrorJSONResponse(w, http.StatusBadRequest, NotMarshall, err.Error())
+		return
+	}
+
+	w.Write(data)
+}
